Address review comments.
